% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manSASInvariance-package.R
\docType{package}
\name{manSASInvariance-package}
\alias{manSASInvariance-package}
\alias{_PACKAGE}
\title{manSASInvariance: Longitudinal Invariance of the Social Axioms Survey}
\description{
Longitudinal invariance of the Social Axioms Survey (SAS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sigmaresearch100/manSASInvariance}
  \item \url{https://sigmaresearch100.github.io/manSASInvariance/}
  \item Report bugs at \url{https://github.com/sigmaresearch100/manSASInvariance/issues}
}

}
\author{
\strong{Maintainer}: Sigma Research \email{r.sigmaresearch100@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID}) [copyright holder]

}
\keyword{internal}
